requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
mysqlclient==2.0.3
PyJWT==2.1.0
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
